fixed whatever I messed up last time

+++ genotypeCells.py
@@ -11,7 +11,7 @@ barcodeToGeneFile = sys.argv[3]
 readThreshFrac = float(sys.argv[4])
 umiThreshFrac = float(sys.argv[5])
 
-minUMIFrac = int(sys.argv[6])
+minUMIFrac = float(sys.argv[6])
 
 with open(barcodeFile) as f:
     cellBarcodes = cp.load(f)
